Lookup key for file to locate .map
After creating my script, rev.manifest function was trying to merge files into the manifest as the same time, so the process ended up being moved to the end once all files had been generated. I could have also put revisions on the .map files so the filenames would match, but I'd have also needed to change the comments within the files. It was easier to leave the .map files without the cache busting revisions element as if a developer needs the maps, cache usually disabled (and not sure if browser cache the map files). Therefore, to have this option to delete matching .map files, I've updated the script to check the original filename of a revised file and check if that file exists with a map extension to auto-delete it.